Big restructuring off the project files
