+++ src/task-manager-app-backend/Abstractions/IHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace task_manager_app_backend.Abstractions
+{
+  public interface IHandler
+  {
+  }
+}
