--- src/task-manager-app-backend/Areas/Users/LoginHandler.cs
@@ -1,56 +1,0 @@
-using Microsoft.AspNetCore.Identity;
-using Microsoft.IdentityModel.Tokens;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text;
-using task_manager_app_backend.Authentication;
-using task_manager_app_backend.Services.Abstractions;
-namespace task_manager_app_backend.Areas.Users;
-
-public class LoginHandler : IHandler
-{
-  private readonly UserManager<ApplicationUser> _userManager;
-  private readonly IConfiguration _configuration;
-
-  public LoginHandler(UserManager<ApplicationUser> userManager, IConfiguration configuration)
-  {
-    _userManager = userManager;
-    _configuration = configuration;
-  }
-
-  public async Task<JwtSecurityToken> LoginUserAsync(LoginModel model)
-  {
-    var user = await _userManager.FindByNameAsync(model.Username);
-    if (user != null && await _userManager.CheckPasswordAsync(user, model.Password))
-    {
-      var userRoles = await _userManager.GetRolesAsync(user);
-
-      var authClaims = new List<Claim>
-            {
-                new Claim(ClaimTypes.Name, user.UserName),
-                new Claim(ClaimTypes.Email, user.Email),
-                new Claim(ClaimTypes.NameIdentifier, user.Id),
-                new Claim(ClaimTypes.Role, userRoles.FirstOrDefault() ?? string.Empty),
-                new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
-            };
-
-      foreach (var userRole in userRoles)
-      {
-        authClaims.Add(new Claim(ClaimTypes.Role, userRole));
-      }
-
-      var authSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_configuration["JWT:Secret"]));
-
-      var token = new JwtSecurityToken(
-          issuer: _configuration["JWT:ValidIssuer"],
-          audience: _configuration["JWT:ValidAudience"],
-          expires: DateTime.Now.AddHours(3),
-          claims: authClaims,
-          signingCredentials: new SigningCredentials(authSigningKey, SecurityAlgorithms.HmacSha256)
-          );
-
-      return token;
-    }
-    throw new Exception();
-  }
-}
