+++ src/task-manager-app-backend/Areas/Users/Models/RegisterRequest.cs
@@ -1,0 +1,18 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace task_manager_app_backend.Areas.Users.Models
+{
+    public class RegisterRequest
+    {
+        [Required(ErrorMessage = "User Name is required")]
+        public string Username { get; set; }
+
+        [EmailAddress]
+        [Required(ErrorMessage = "Email is required")]
+        public string Email { get; set; }
+
+        [Required(ErrorMessage = "Password is required")]
+        public string Password { get; set; }
+
+    }
+}
