+++ src/task-manager-app-backend/Areas/Users/Models/Response.cs
@@ -1,0 +1,8 @@
+﻿namespace task_manager_app_backend.Areas.Users.Models
+{
+    public class Response
+    {
+        public string Status { get; set; }
+        public string Message { get; set; }
+    }
+}
