+++ src/task-manager-app-backend/Areas/Users/Services/ApplicationUser.cs
@@ -1,0 +1,8 @@
+﻿using Microsoft.AspNetCore.Identity;
+
+namespace task_manager_app_backend.Areas.Users.Services
+{
+    public class ApplicationUser : IdentityUser
+    {
+    }
+}
