+++ src/task-manager-app-backend/Areas/Users/Services/UserRoles.cs
@@ -1,0 +1,8 @@
+﻿namespace task_manager_app_backend.Areas.Users.Services
+{
+    public static class UserRoles
+    {
+        public const string Admin = "Admin";
+        public const string User = "User";
+    }
+}
