--- src/task-manager-app-backend/Authentication/ApplicationUser.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace task_manager_app_backend.Authentication
-{
-  public class ApplicationUser : IdentityUser
-  {
-  }
-}
