--- src/task-manager-app-backend/Authentication/LoginModel.cs
@@ -1,13 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace task_manager_app_backend.Authentication
-{
-  public class LoginModel
-  {
-    [Required(ErrorMessage = "User Name is required")]
-    public string Username { get; set; }
-
-    [Required(ErrorMessage = "Password is required")]
-    public string Password { get; set; }
-  }
-}
