--- src/task-manager-app-backend/Authentication/RegisterModel.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace task_manager_app_backend.Authentication
-{
-  public class RegisterModel
-  {
-    [Required(ErrorMessage = "User Name is required")]
-    public string Username { get; set; }
-
-    [EmailAddress]
-    [Required(ErrorMessage = "Email is required")]
-    public string Email { get; set; }
-
-    [Required(ErrorMessage = "Password is required")]
-    public string Password { get; set; }
-
-  }
-}
