--- src/task-manager-app-backend/Authentication/Response.cs
@@ -1,8 +1,0 @@
-﻿namespace task_manager_app_backend.Authentication
-{
-  public class Response
-  {
-    public string Status { get; set; }
-    public string Message { get; set; }
-  }
-}
