--- src/task-manager-app-backend/Authentication/UserRoles.cs
@@ -1,8 +1,0 @@
-﻿namespace task_manager_app_backend.Authentication
-{
-  public static class UserRoles
-  {
-    public const string Admin = "Admin";
-    public const string User = "User";
-  }
-}
