+++ src/task-manager-app-backend/Models/AssignmentType.cs
@@ -4,10 +4,10 @@ namespace task_manager_app_backend.Models;
 
 public class AssignmentType
 {
-  public int Id { get; set; }
-  public string Name { get; set; }
-  [JsonIgnore]
-  public ICollection<Assignment> Assignments { get; set;}
+    public int Id { get; set; }
+    public string Name { get; set; }
+    [JsonIgnore]
+    public ICollection<Assignment> Assignments { get; set; }
 
 }
