--- src/task-manager-app-backend/Services/Abstractions/IHandler.cs
@@ -1,6 +1,0 @@
-﻿namespace task_manager_app_backend.Services.Abstractions
-{
-  public interface IHandler
-  {
-  }
-}
