+++ src/task-manager-app-backend/appsettings.json
@@ -7,7 +7,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost\\SQLEXPRESS; Database=TaskManagerDatabase;Trusted_Connection=True;TrustServerCertificate=True;"
+    "DefaultConnection": "Server=localhost\\SQLEXPRESS01; Database=TaskManagerDatabase;Trusted_Connection=True;TrustServerCertificate=True;"
   },
   "JWT": {
     "ValidAudience": "http://localhost:5104",
