+++ src/task-manager-app-backend/task-manager-app-backend.csproj
@@ -27,4 +27,9 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Areas\Users\Controllers\" />
+    <Folder Include="Authentication\" />
+  </ItemGroup>
+
 </Project>
